chore: move helper to separate package
