+++ aws-lambda-router/internal/router.go
@@ -3,6 +3,7 @@ package internal
 import (
 	"fmt"
 	"github.com/wundergraph/cosmo/router/core"
+	"github.com/wundergraph/cosmo/router/pkg/execution_config"
 	"github.com/wundergraph/cosmo/router/pkg/metric"
 	"github.com/wundergraph/cosmo/router/pkg/trace"
 	"go.uber.org/zap"
@@ -36,7 +37,7 @@ func NewRouter(opts ...Option) *core.Router {
 
 	logger := rc.Logger
 
-	routerConfig, err := core.SerializeConfigFromFile(rc.RouterConfigPath)
+	routerConfig, err := execution_config.SerializeConfigFromFile(rc.RouterConfigPath)
 	if err != nil {
 		logger.Fatal("Could not read router config", zap.Error(err), zap.String("path", rc.RouterConfigPath))
 	}
