+++ go.work.sum
@@ -190,6 +190,7 @@ github.com/urfave/cli/v2 v2.8.1/go.mod h1:Z41J9TPoffeoqP0Iza0YbAhGvymRdZAd2uPmZ5
 github.com/urfave/cli/v2 v2.25.5/go.mod h1:GHupkWPMM0M/sj1a2b4wUrWBPzazNrIjouW6fmdJLxc=
 github.com/vektah/gqlparser/v2 v2.5.1/go.mod h1:mPgqFBu/woKTVYWyNk8cO3kh4S/f4aRFZrvOnp3hmCs=
 github.com/vektah/gqlparser/v2 v2.5.10/go.mod h1:1rCcfwB2ekJofmluGWXMSEnPMZgbxzwj6FaZ/4OT8Cc=
+github.com/wundergraph/cosmo/demo v0.0.0-20240430083233-f6907a5c2a0a/go.mod h1:mVUJqOK+CUeQZ3ofbXM4I/TgSt39DXiHcZollOAX6Ws=
 github.com/wundergraph/cosmo/router v0.0.0-20240403095627-846af216866c/go.mod h1:LS+5qlr4fQVEW7JMXXI1sz7CH5cdnqx3BNc10p+UbW4=
 github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.11/go.mod h1:26OPLppZhb4zc9RCfuLPn1Zfv3hNs/fiN4U6d9pLOw4=
 github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.13/go.mod h1:e71oBiNorTfjTi6eDRCFoIoJiFWcwiQ0eObm75VzJh0=
@@ -197,8 +198,6 @@ github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.13.0.20240319131455-d4f2d0b
 github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.13.0.20240319131455-d4f2d0b32791/go.mod h1:e71oBiNorTfjTi6eDRCFoIoJiFWcwiQ0eObm75VzJh0=
 github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.17 h1:xw4RZNahoZoWj5jJXi7XYvznTXiPjYkb8iGvFa4p+vc=
 github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.17/go.mod h1:e71oBiNorTfjTi6eDRCFoIoJiFWcwiQ0eObm75VzJh0=
-github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.31 h1:ctq+CnvXkbJQvduxkl0yrCn4wAqseUAFMud8cE+XxRM=
-github.com/wundergraph/graphql-go-tools/v2 v2.0.0-rc.31/go.mod h1:hNR2C7S1M+c9Ap24tHCEMe9gFY9K3smX46x5E1U1NQw=
 github.com/xdg-go/pbkdf2 v1.0.0 h1:Su7DPu48wXMwC3bs7MCNG+z4FhcyEuz5dlvchbq0B0c=
 github.com/xdg-go/scram v1.1.1 h1:VOMT+81stJgXW3CpHyqHN3AXDYIMsx56mEFrB37Mb/E=
 github.com/xdg-go/stringprep v1.0.3 h1:kdwGpVNwPFtjs98xCGkHjQtGKh86rDcRZN17QEMCOIs=
